+++ src/main/java/my/comunity/common/controller/QuestionControll.java
@@ -13,7 +13,7 @@ public class QuestionControll {
     @Autowired
     QuestionService questionMapper;
     @GetMapping("question/{id}")
-    private String question(@PathVariable("id") Integer id, Model model){
+    private String question(@PathVariable("id") Long id, Model model){
         QuestionDto questionDto=questionMapper.getById(id);
         model.addAttribute("question",questionDto);
         model.addAttribute("info","ceshi");
